move tileset to own file + remove unused functions
Moved Tileset classes to a separate file to clean up header and remove some functions that were not in use
