+++ source/gamemap/TileContainer.cpp
@@ -323,6 +323,11 @@ private:
     double mHiddenValueSouth;
 };
 
+bool sortByDistSquared(const TileDistance& tileDist1, const TileDistance& tileDist2)
+{
+    return tileDist1.getDistSquared() < tileDist2.getDistSquared();
+}
+
 TileContainer::TileContainer(int initTileDistance):
     mMapSizeX(0),
     mMapSizeY(0),
@@ -332,76 +337,6 @@ TileContainer::TileContainer(int initTileDistance):
 {
     buildTileDistance(initTileDistance);
 }
-
-TileSet::TileSet(const Ogre::Vector3& scale) :
-    mTileValues(static_cast<uint32_t>(TileVisual::countTileVisual), std::vector<TileSetValue>(16)),
-    mScale(scale),
-    mTileLinks(std::vector<uint32_t>(static_cast<uint32_t>(TileVisual::countTileVisual), 0))
-{
-}
-
-std::vector<TileSetValue>& TileSet::configureTileValues(TileVisual tileVisual)
-{
-    uint32_t tileTypeNumber = static_cast<uint32_t>(tileVisual);
-    if(tileTypeNumber >= mTileValues.size())
-    {
-        OD_ASSERT_TRUE_MSG(false, "Trying to get unknow tileType=" + Tile::tileVisualToString(tileVisual));
-        return mTileValues.at(0);
-    }
-
-    return mTileValues[tileTypeNumber];
-}
-
-const std::vector<TileSetValue>& TileSet::getTileValues(TileVisual tileVisual) const
-{
-    uint32_t tileTypeNumber = static_cast<uint32_t>(tileVisual);
-    if(tileTypeNumber >= mTileValues.size())
-    {
-        OD_ASSERT_TRUE_MSG(false, "Trying to get unknow tileType=" + Tile::tileVisualToString(tileVisual));
-        return mTileValues.at(0);
-    }
-
-    return mTileValues[tileTypeNumber];
-}
-
-void TileSet::addTileLink(TileVisual tileVisual1, TileVisual tileVisual2)
-{
-    uint32_t intTile1Visual = static_cast<uint32_t>(tileVisual1);
-    uint32_t intTile2Visual = static_cast<uint32_t>(tileVisual2);
-    if(intTile1Visual >= mTileLinks.size())
-    {
-        OD_ASSERT_TRUE_MSG(false, "TileVisual=" + Tile::tileVisualToString(tileVisual1));
-        return;
-    }
-    if(intTile2Visual >= mTileLinks.size())
-    {
-        OD_ASSERT_TRUE_MSG(false, "TileVisual=" + Tile::tileVisualToString(tileVisual2));
-        return;
-    }
-
-    uint32_t tileLink;
-    tileLink = 1 << intTile2Visual;
-    mTileLinks[intTile1Visual] |= tileLink;
-
-    // We want a commutative relationship
-    tileLink = 1 << intTile1Visual;
-    mTileLinks[intTile2Visual] |= tileLink;
-}
-
-bool TileSet::areLinked(const Tile* tile1, const Tile* tile2) const
-{
-    // Check if the tile visual is linkable
-    uint32_t intTile1Visual = static_cast<uint32_t>(tile1->getTileVisual());
-    uint32_t intTile2Visual = static_cast<uint32_t>(tile2->getTileVisual());
-    uint32_t linkValue = 1 << intTile2Visual;
-    linkValue &= mTileLinks[intTile1Visual];
-
-    if(linkValue == 0)
-        return false;
-
-    return true;
-}
-
 
 TileContainer::~TileContainer()
 {
@@ -497,11 +432,6 @@ Tile* TileContainer::tileFromPacket(ODPacket& packet) const
     return tile;
 }
 
-unsigned int TileContainer::numTiles()
-{
-    return mMapSizeX * mMapSizeY;
-}
-
 bool TileContainer::allocateMapMemory(int xSize, int ySize)
 {
     if (xSize <= 0 || ySize <= 0)
@@ -544,16 +474,6 @@ bool TileContainer::allocateMapMemory(int xSize, int ySize)
     }
 
     return true;
-}
-
-TileType TileContainer::getSafeTileType(const Tile* tt) const
-{
-    return (tt == nullptr) ? TileType::nullTileType : tt->getType();
-}
-
-bool TileContainer::getSafeTileFullness(const Tile* tt) const
-{
-    return (tt == nullptr) ? false : (tt->getFullness() > 0);
 }
 
 std::vector<Tile*> TileContainer::rectangularRegion(int x1, int y1, int x2, int y2)
@@ -796,11 +716,6 @@ void TileContainer::buildTileDistance(int distance)
     }
 
     mTileDistanceComputed = distance;
-}
-
-bool TileContainer::sortByDistSquared(const TileDistance& tileDist1, const TileDistance& tileDist2)
-{
-    return tileDist1.getDistSquared() < tileDist2.getDistSquared();
 }
 
 std::list<Tile*> TileContainer::tilesBetween(int x1, int y1, int x2, int y2) const
